+++ addy/AppShortcutsProvider/AddyShortcuts.swift
@@ -15,11 +15,25 @@ struct AddyShortcuts: AppShortcutsProvider {
                     phrases: [
                         "Add a new alias in \(.applicationName)",
                         "Add an alias in \(.applicationName)",
+                        "Add alias in \(.applicationName)",
                         "Create a new alias in \(.applicationName)",
                         "Create an alias in \(.applicationName)",
+                        "Create alias in \(.applicationName)",
                         "Generate a new alias in \(.applicationName)",
                     ],
                     shortTitle: "app_intent_add_alias",
                     systemImageName: "plus")
+
+        AppShortcut(intent: FindAliasIntent(),
+                    phrases: [
+                        "Find an alias in \(.applicationName)",
+                        "Find alias in \(.applicationName)",
+                        "Search alias in \(.applicationName)",
+                        "Search for an alias in \(.applicationName)",
+                        "Look up an alias in \(.applicationName)",
+                        "Get an alias in \(.applicationName)",
+                    ],
+                    shortTitle: "app_intent_find_alias",
+                    systemImageName: "magnifyingglass")
     }
 }